Split admin controller into settings and onboarding controllers
